curriculum/tracks/databricks/project_01_news_classifier/pyproject.toml
[project]
name = "news-classifier-agent"
version = "0.1.0"
description = "MLflow experiments for news classification using external and internal LLMs"
requires-python = ">=3.10"
dependencies = [
    "mlflow[databricks]>=3.1",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "databricks-sdk>=0.18.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"